# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "V-LAB PROJECT DOCUMENTATION",
        "contact": {
            "email": "man2@cin.ufpe.br"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/category/{id_user}": {
            "get": {
                "tags": [
                    "Categoria"
                ],
                "summary": "Listar todas as Categorias do Usuário Especificado",
                "description": "Usuário pode utilizar para visualizar todas as suas categorias.",
                "operationId": "categoryIndex",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação Bem Seucedida"
                    },
                    "404": {
                        "description": "Nenhuma Categoria Encontrado Para Esse Usuário"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categoria"
                ],
                "summary": "Cadastrar Nova Categoria Para o Usuário Especificado",
                "description": "Usuário pode utilizar para cadastrar uma nova categoria na sua conta.",
                "operationId": "categoryStore",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados da Categoria",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Input Inválido"
                    },
                    "409": {
                        "description": "Categoria Já Existe"
                    },
                    "201": {
                        "description": "Categoria Adicionada no Banco de Dados"
                    },
                    "500": {
                        "description": "Algo Errado no Insert"
                    }
                }
            }
        },
        "/api/category/{id_user}/delete/{id_category}": {
            "delete": {
                "tags": [
                    "Categoria"
                ],
                "summary": "Deletar a Categoria Especificada do Usuário Especificado",
                "description": "Usuário pode utilizar para deletar uma categoria da sua conta. Todas as transações dessa categoria apagada são alteradas para categorial 'Geral'.",
                "operationId": "categoryDestroy",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id_category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Acesso Negado. Categoria 'Geral' não pode ser excluída."
                    },
                    "200": {
                        "description": "Categoria Excluída e transações atualizadas."
                    },
                    "404": {
                        "description": "Categoria não Encontrado"
                    }
                }
            }
        },
...
    }
}